Extract common test code into an internal library
