+++ Funcky.Test/Extensions/EnumerableExtensions/AverageOrNoneTest.cs
@@ -1,5 +1,6 @@
 using FsCheck;
 using FsCheck.Xunit;
+using Funcky.Test.Internal;
 using Funcky.Test.TestUtils;
 
 namespace Funcky.Test.Extensions.EnumerableExtensions
