+++ Funcky.Test/Extensions/EnumerableExtensions/MinOrNoneTest.cs
@@ -1,5 +1,7 @@
 using FsCheck;
 using FsCheck.Xunit;
+using Funcky.Test.Internal;
+using Funcky.Test.Internal.Data;
 using Funcky.Test.TestUtils;
 
 namespace Funcky.Test.Extensions.EnumerableExtensions
